http2: specific file for decompression
